+++ src/main/java/com/smart/house/app/service/SmartDeviceService.java
@@ -6,10 +6,13 @@ import com.smart.house.app.dto.material.MaterialRequestDto;
 import com.smart.house.app.dto.material.MaterialResponseDto;
 import com.smart.house.app.entity.Material;
 import com.smart.house.app.entity.SmartDevice;
+import com.smart.house.app.exception.CustomEntityNotFoundException;
 import com.smart.house.app.repository.DeviceRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,16 +21,19 @@ public class SmartDeviceService {
     private final DeviceRepository deviceRepository;
 
 
-    public SmartDeviceResponseDto getSmartDevice(String name){
-        SmartDevice smartDeviceEntity = deviceRepository.findByName(name).orElseThrow(() ->
-                new EntityNotFoundException("Device not found"));
-        return SmartDeviceResponseDto.builder()
-                .name(smartDeviceEntity.getName())
-                .build();
-
+    public SmartDeviceResponseDto getSmartDevice(String name) throws CustomEntityNotFoundException {
+        Optional<SmartDevice> smartDeviceOptional = deviceRepository.findByName(name);
+        if (smartDeviceOptional.isPresent()) {
+            SmartDevice smartDeviceEntity = smartDeviceOptional.get();
+            return SmartDeviceResponseDto.builder()
+                    .name(smartDeviceEntity.getName())
+                    .build();
+        } else {
+            throw new CustomEntityNotFoundException("Smart device not found");
+        }
     }
 
-    public SmartDeviceResponseDto createSmartDevice(SmartDeviceRequestDto smartDeviceRequestDto){
+    public SmartDeviceResponseDto createSmartDevice(SmartDeviceRequestDto smartDeviceRequestDto) {
         SmartDevice smartDeviceEntity = new SmartDevice();
         smartDeviceEntity.setName(smartDeviceRequestDto.getName());
         smartDeviceEntity.setSmartDeviceType(smartDeviceRequestDto.getSmartDeviceType());
@@ -38,10 +44,13 @@ public class SmartDeviceService {
     }
 
 
-
-    public void deleteSmartDevice(Long id) {
-        SmartDevice smartDeviceEntity = deviceRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("Device not found"));
-        deviceRepository.delete(smartDeviceEntity);
+    public void deleteSmartDevice(Long id) throws CustomEntityNotFoundException {
+        Optional<SmartDevice> smartDeviceOptional = deviceRepository.findById(id);
+        if (smartDeviceOptional.isPresent()) {
+            SmartDevice smartDeviceEntity = smartDeviceOptional.get();
+            deviceRepository.delete(smartDeviceEntity);
+        } else {
+            throw new CustomEntityNotFoundException("Smart device not found");
+        }
     }
 }
