+++ src/main/java/com/smart/house/app/service/SmartDeviceTypeService.java
@@ -6,10 +6,13 @@ import com.smart.house.app.dto.user.UserRequestDto;
 import com.smart.house.app.dto.user.UserResponseDto;
 import com.smart.house.app.entity.SmartDeviceType;
 import com.smart.house.app.entity.User;
+import com.smart.house.app.exception.CustomEntityNotFoundException;
 import com.smart.house.app.repository.SmartDeviceTypeRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,16 +20,20 @@ public class SmartDeviceTypeService {
 
     private final SmartDeviceTypeRepository smartDeviceTypeRepository;
 
-    public SmartDeviceTypeResponseDto getSmartDeviceType(Long id){
-        SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("smartDeviceType not found"));
-        return SmartDeviceTypeResponseDto.builder()
-                .type(smartDeviceTypeEntity.getType())
-                .build();
+    public SmartDeviceTypeResponseDto getSmartDeviceType(Long id) throws CustomEntityNotFoundException {
+        Optional<SmartDeviceType> smartDeviceTypeOptional = smartDeviceTypeRepository.findById(id);
+        if (smartDeviceTypeOptional.isPresent()) {
+            SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeOptional.get();
+            return SmartDeviceTypeResponseDto.builder()
+                    .type(smartDeviceTypeEntity.getType())
+                    .build();
+        } else {
+            throw new CustomEntityNotFoundException("Smart device type not found");
+        }
     }
 
 
-    public SmartDeviceTypeResponseDto createSmartDeviceType(SmartDeviceTypeRequestDto smartDeviceTypeRequestDto){
+    public SmartDeviceTypeResponseDto createSmartDeviceType(SmartDeviceTypeRequestDto smartDeviceTypeRequestDto) {
         SmartDeviceType smartDeviceTypeEntity = new SmartDeviceType();
         smartDeviceTypeEntity.setType(smartDeviceTypeRequestDto.getType());
         SmartDeviceType result = smartDeviceTypeRepository.save(smartDeviceTypeEntity);
@@ -35,19 +42,28 @@ public class SmartDeviceTypeService {
                 .build();
     }
 
-    public SmartDeviceTypeResponseDto changeSmartDeviceType(SmartDeviceTypeRequestDto smartDeviceTypeRequestDto, Long id){
-        SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("smartDeviceType not found"));
-        smartDeviceTypeEntity.setType(smartDeviceTypeRequestDto.getType());
-        SmartDeviceType result = smartDeviceTypeRepository.save(smartDeviceTypeEntity);
-        return SmartDeviceTypeResponseDto.builder()
-                .type(smartDeviceTypeEntity.getType())
-                .build();
+    public SmartDeviceTypeResponseDto changeSmartDeviceType(SmartDeviceTypeRequestDto smartDeviceTypeRequestDto, Long id) throws CustomEntityNotFoundException {
+        Optional<SmartDeviceType> smartDeviceTypeOptional = smartDeviceTypeRepository.findById(id);
+        if (smartDeviceTypeOptional.isPresent()) {
+            SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeOptional.get();
+            smartDeviceTypeEntity.setType(smartDeviceTypeRequestDto.getType());
+            SmartDeviceType result = smartDeviceTypeRepository.save(smartDeviceTypeEntity);
+            return SmartDeviceTypeResponseDto.builder()
+                    .type(smartDeviceTypeEntity.getType())
+                    .build();
+        } else {
+            throw new CustomEntityNotFoundException("Smart device type not found");
+        }
     }
 
-    public void deleteSmartDeviceType(Long id){
-        SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("smartDeviceType not found"));
-        smartDeviceTypeRepository.delete(smartDeviceTypeEntity);
+    public void deleteSmartDeviceType(Long id) throws CustomEntityNotFoundException {
+        Optional<SmartDeviceType> smartDeviceTypeOptional = smartDeviceTypeRepository.findById(id);
+        if (smartDeviceTypeOptional.isPresent()) {
+            SmartDeviceType smartDeviceTypeEntity = smartDeviceTypeOptional.get();
+            smartDeviceTypeRepository.delete(smartDeviceTypeEntity);
+        } else {
+            throw new CustomEntityNotFoundException("Smart device type not found");
+        }
     }
 }
+
